+++ defeatbeta_api/utils/const.py
@@ -12,9 +12,10 @@ stock_dividend_events="stock_dividend_events"
 stock_split_events="stock_split_events"
 exchange_rate="exchange_rate"
 daily_treasury_yield="daily_treasury_yield"
+stock_earning_call_transcripts="stock_earning_call_transcripts"
 
 tables = [stock_profile, stock_officers, stock_summary, stock_tailing_eps, stock_earning_calendar, stock_revenue_estimates, stock_earning_estimates,
-                stock_historical_eps, stock_statement, stock_prices, stock_dividend_events, stock_split_events, exchange_rate, daily_treasury_yield]
+                stock_historical_eps, stock_statement, stock_prices, stock_dividend_events, stock_split_events, exchange_rate, daily_treasury_yield, stock_earning_call_transcripts]
 
 income_statement = "income_statement"
 balance_sheet = "balance_sheet"