+++ test/test_duckdb_client.py
@@ -17,6 +17,10 @@ class TestDuckDBClient(unittest.TestCase):
             )
             print(result)
             result = client.query(
+                "SELECT symbol,fiscal_year,fiscal_quarter,report_date,unnest(transcripts).paragraph_number as paragraph_number,unnest(transcripts).speaker as speaker,unnest(transcripts).content as content from 'https://huggingface.co/datasets/bwzheng2010/yahoo-finance-data/resolve/main/data/stock_earning_call_transcripts.parquet' where symbol='BABA' and fiscal_year=2025 and fiscal_quarter=2;"
+            )
+            print(result)
+            result = client.query(
                 "SELECT * FROM cache_httpfs_cache_access_info_query()"
             )
             print(result)